fixed obstacle removal and graphics glitches

+++ core/src/com/mygdx/game/objects/Background.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.graphics.Texture;
  * Created by Syuqri on 3/16/2016.
  */
 public class Background extends Movable{
-    public Background(int y){
+    public Background(float y){
         super(0, y, 480, 800);
         this.setImage(new Texture(Gdx.files.internal("bg.png")));
     }
